Remove installer logic from __init__.py and refactor arg parsing to leverage dispatch pattern.
