+++ ext/dcos-installer/dcos_installer/cli.py
@@ -1,108 +1,163 @@
-import asyncio
+import argparse
+import coloredlogs
 import logging
 import sys
 
-from dcos_installer import action_lib, backend
-from dcos_installer.util import CONFIG_PATH
-from dcos_installer.installer_analytics import InstallerAnalytics
-from dcos_installer.action_lib.prettyprint import print_header, PrettyPrint
-from ssh.utils import AbstractSSHLibDelegate
+from dcos_installer.cli_dispatcher import dispatch_option, dispatch_action
 
 
 log = logging.getLogger(__name__)
 
 
-class CliDelegate(AbstractSSHLibDelegate):
-    def on_update(self, future, callback_called):
-        chain_name, result_object, host = future.result()
-        callback_called.set_result(True)
-        log.debug('on_update executed for {}'.format(chain_name))
-
-    def on_done(self, name, result, host_status=None):
-        print_header('STAGE {}'.format(name))
-
-    def prepare_status(self, name, nodes):
-        pass
-
-
-def run_loop(action, options):
-    assert callable(action)
-    loop = asyncio.new_event_loop()
-    asyncio.set_event_loop(loop)
-
-    print_header('START {}'.format(action.__name__))
-    try:
-        config = backend.get_config()
-        cli_delegate = CliDelegate()
-        result = loop.run_until_complete(action(config, block=True, async_delegate=cli_delegate, options=options))
-        pp = PrettyPrint(result)
-        pp.stage_name = action.__name__
-        pp.beautify('print_data')
-
-    finally:
-        loop.close()
-    exitcode = 0
-    for host_result in result:
-        for command_result in host_result:
-            for host, process_result in command_result.items():
-                if process_result['returncode'] != 0:
-                    exitcode += 1
-    print_header('ACTION {} COMPLETE'.format(action.__name__))
-    pp.print_summary()
-    return exitcode
+def setup_logger(options):
+    level = 'INFO'
+    if options.verbose:
+        level = 'DEBUG'
+    coloredlogs.install(
+        level=level,
+        level_styles={
+            'warn': {
+                'color': 'yellow'
+            },
+            'error': {
+                'color': 'red',
+                'bold': True,
+            },
+        },
+        fmt='%(message)s',
+        isatty=True
+    )
+    log.debug("Logger set to DEBUG")
 
 
-def tall_enough_to_ride():
-    choices_true = ['yes', 'y']
-    choices_false = ['no', 'n']
-    while True:
-        do_uninstall = input('This will uninstall DC/OS on your cluster. You may need to manually remove '
-                             '/var/lib/zookeeper in some cases after this completes, please see our documentation '
-                             'for details. Are you ABSOLUTELY sure you want to proceed? [ (y)es/(n)o ]: ')
-        if do_uninstall.lower() in choices_true:
-            return
-        elif do_uninstall.lower() in choices_false:
-            sys.exit(1)
-        else:
-            log.error('Choices are [y]es or [n]o. "{}" is not a choice'.format(do_uninstall))
+def start_installer(args):
+    """
+    The web based installer leverages Flask to present end-users of
+    dcos_installer with a clean web interface to configure their
+    site-based installation of DC/OS.
+    """
+    # If no args are passed to the class, then we're calling this
+    # class from another library or code so we shouldn't execute
+    # parser or anything else
+    options = parse_args(args)
+    setup_logger(options)
+    dispatch_option(options)
+    dispatch_action(options)
 
 
-def get_new_analytics(config_path=CONFIG_PATH):
-    return InstallerAnalytics(backend.get_config(config_path))
+def parse_args(args):
+    def print_usage():
+        return """
+Install Mesosophere's Data Center Operating System
+
+dcos_installer [-h] [-f LOG_FILE] [--hash-password HASH_PASSWORD] [-v]
+[--web | --genconf | --preflight | --deploy | --postflight | --uninstall | --validate-config | --test]
+
+Environment Settings:
+
+PORT                  Set the :port to run the web UI
+CHANNEL_NAME          ADVANCED - Set build channel name
+BOOTSTRAP_ID          ADVANCED - Set bootstrap ID for build
+
+"""
+
+    """
+    Parse CLI arguments and return a map of options.
+    """
+    parser = argparse.ArgumentParser(usage=print_usage())
+    mutual_exc = parser.add_mutually_exclusive_group()
+
+    parser.add_argument(
+        '--hash-password',
+        default='',
+        type=str,
+        help='Hash a password on the CLI for use in the config.yaml.'
+    )
+
+    parser.add_argument(
+        '-v',
+        '--verbose',
+        action='store_true',
+        default=False,
+        help='Verbose log output (DEBUG).')
+
+    parser.add_argument(
+        '-p',
+        '--port',
+        type=int,
+        default=9000,
+        help=argparse.SUPPRESS)
+
+    parser.add_argument(
+        '--offline',
+        action='store_true',
+        default=False,
+        help='Do not install preflight prerequisites on CentOS7, RHEL7 in web mode'
+    )
+
+    mutual_exc.add_argument(
+        '--web',
+        action='store_true',
+        default=False,
+        help='Run the web interface.')
+
+    mutual_exc.add_argument(
+        '--genconf',
+        action='store_true',
+        default=False,
+        help='Execute the configuration generation (genconf).')
+
+    mutual_exc.add_argument(
+        '--preflight',
+        action='store_true',
+        default=False,
+        help='Execute the preflight checks on a series of nodes.')
+
+    mutual_exc.add_argument(
+        '--install-prereqs',
+        action='store_true',
+        default=False,
+        help='Install preflight prerequisites. Works only on CentOS7 and RHEL7.')
+
+    mutual_exc.add_argument(
+        '--deploy',
+        action='store_true',
+        default=False,
+        help='Execute a deploy.')
+
+    mutual_exc.add_argument(
+        '--postflight',
+        action='store_true',
+        default=False,
+        help='Execute postflight checks on a series of nodes.')
+
+    mutual_exc.add_argument(
+        '--uninstall',
+        action='store_true',
+        default=False,
+        help='Execute uninstall on target hosts.')
+
+    mutual_exc.add_argument(
+        '--validate-config',
+        action='store_true',
+        default=False,
+        help='Validate the configuration for executing --genconf and deploy arguments in config.yaml')
+
+    mutual_exc.add_argument(
+        '--test',
+        action='store_true',
+        default=False,
+        help='Performs tests on the dcos_installer application')
+
+    options = parser.parse_args(args)
+    return options
 
 
-def dispatch_action(args):
-    installer_analytics = get_new_analytics()
+def main():
+    if len(sys.argv) == 1:
+        start_installer(["--genconf"])
+    else:
+        start_installer(sys.argv[1:])
 
-    def preflight(args):
-        print_header("EXECUTING PREFLIGHT")
-        return action_lib.run_preflight
-
-    def postflight(args):
-        print_header("EXECUTING POSTFLIGHT")
-        return action_lib.run_postflight
-
-    def deploy(args):
-        print_header("EXECUTING DC/OS INSTALLATION")
-        return action_lib.install_dcos
-
-    def uninstall(args):
-        print_header("EXECUTING UNINSTALL")
-        tall_enough_to_ride()
-        return action_lib.uninstall_dcos
-
-    def install_prereqs(args):
-        print_header("EXECUTING INSTALL PREREQUISITES")
-        return action_lib.install_prereqs
-
-    for action in ['deploy', 'preflight', 'postflight', 'uninstall', 'install_prereqs']:
-        if getattr(args, action):
-            errors = run_loop(locals()[action](args), args)
-            if not args.disable_analytics:
-                installer_analytics.send(
-                    action=action,
-                    install_method="cli",
-                    num_errors=errors,
-                )
-            exit_code = 1 if errors > 0 else 0
-            sys.exit(exit_code)
+if __name__ == '__main__':
+    main()
